# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InvoiceFlow API",
    "version": "1.0.0",
    "description": "Money is always integer minor units. Authenticated routes need a Supabase access token. Cross-business ids answer 404. Generated by api/scripts/build-docs.ts."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "supabaseJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/payments/webhook": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Stripe webhook (signature-verified, raw body)",
        "operationId": "post_v1_payments_webhook",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/pay/{token}": {
      "get": {
        "tags": ["Public"],
        "summary": "Customer-facing invoice and pay page (HTML)",
        "operationId": "get_pay_token",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/public/invoices/{token}": {
      "get": {
        "tags": ["Public"],
        "summary": "Sanitised invoice JSON",
        "operationId": "get_public_invoices_token",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/public/invoices/{token}/pdf": {
      "get": {
        "tags": ["Public"],
        "summary": "Invoice PDF",
        "operationId": "get_public_invoices_token_pdf",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/public/invoices/{token}/view": {
      "post": {
        "tags": ["Public"],
        "summary": "Record that the customer viewed the invoice",
        "operationId": "post_public_invoices_token_view",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/public/invoices/{token}/payment-intent": {
      "post": {
        "tags": ["Public"],
        "summary": "Start a payment from the public page",
        "operationId": "post_public_invoices_token_payment_intent",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/estimate/{token}": {
      "get": {
        "tags": ["Public"],
        "summary": "Customer-facing estimate page (HTML) with Accept / Decline",
        "operationId": "get_estimate_token",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/public/estimates/{token}": {
      "get": {
        "tags": ["Public"],
        "summary": "Sanitised estimate JSON",
        "operationId": "get_public_estimates_token",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/public/estimates/{token}/pdf": {
      "get": {
        "tags": ["Public"],
        "summary": "Estimate PDF",
        "operationId": "get_public_estimates_token_pdf",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/public/estimates/{token}/view": {
      "post": {
        "tags": ["Public"],
        "summary": "Record that the customer viewed the estimate",
        "operationId": "post_public_estimates_token_view",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/public/estimates/{token}/respond": {
      "post": {
        "tags": ["Public"],
        "summary": "Customer accepts or declines ({decision: accept|decline, name?}); the same answer twice is a no-op, the opposite answer is 409",
        "operationId": "post_public_estimates_token_respond",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/stripe/connect/return": {
      "get": {
        "tags": ["Public"],
        "summary": "Stripe onboarding return page",
        "operationId": "get_stripe_connect_return",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/stripe/connect/refresh": {
      "get": {
        "tags": ["Public"],
        "summary": "Stripe onboarding refresh page",
        "operationId": "get_stripe_connect_refresh",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    },
    "/.well-known/apple-developer-merchantid-domain-association": {
      "get": {
        "tags": ["Public"],
        "summary": "Apple Pay domain verification file",
        "operationId": "get_well_known_apple_developer_merchantid_domain_association",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "security": []
      }
    }
  }
}